# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceCircle API",
    "description": "Backend for VoiceCircle: voice enrollment, practice scam calls, deepfake checks and Hugh companion calls.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/dev/calls/{call_id}/say": {
      "post": {
        "tags": [
          "simulator"
        ],
        "summary": "Sim Say",
        "description": "Mock mode: pretend the senior said something on the call.",
        "operationId": "sim_say_api_v1_dev_calls__call_id__say_post",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimSayIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dev/calls/{call_id}/hangup": {
      "post": {
        "tags": [
          "simulator"
        ],
        "summary": "Sim Hangup",
        "description": "Mock mode: pretend the senior hung up.",
        "operationId": "sim_hangup_api_v1_dev_calls__call_id__hangup_post",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dev/sms": {
      "get": {
        "tags": [
          "simulator"
        ],
        "summary": "Sim Sms",
        "description": "Mock mode: SMS messages that would have been sent.",
        "operationId": "sim_sms_api_v1_dev_sms_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/seed": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Demo Seed",
        "description": "Create a fully populated demo circle for the current user (DEMO_MODE only).",
        "operationId": "demo_seed_api_v1_demo_seed_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/reset": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Demo Reset",
        "description": "Delete every circle owned by the current user (DEMO_MODE only).",
        "operationId": "demo_reset_api_v1_demo_reset_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CircleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "Members"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CircleStats"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id",
          "created_at",
          "members"
        ],
        "title": "CircleOut"
      },
      "CircleStats": {
        "properties": {
          "voices_enrolled": {
            "type": "integer",
            "title": "Voices Enrolled"
          },
          "last_practice_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Practice Score"
          },
          "last_practice_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Practice At"
          },
          "last_companion_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Companion At"
          },
          "open_alerts": {
            "type": "integer",
            "title": "Open Alerts"
          },
          "detections": {
            "type": "integer",
            "title": "Detections"
          }
        },
        "type": "object",
        "required": [
          "voices_enrolled",
          "last_practice_score",
          "last_practice_at",
          "last_companion_at",
          "open_alerts",
          "detections"
        ],
        "title": "CircleStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "circle_id": {
            "type": "string",
            "title": "Circle Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "phone_e164": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone E164"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "companion_call_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companion Call Time"
          },
          "companion_enabled": {
            "type": "boolean",
            "title": "Companion Enabled"
          },
          "personal_facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Personal Facts",
            "default": []
          },
          "voice_status": {
            "type": "string",
            "title": "Voice Status",
            "default": "none"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "circle_id",
          "user_id",
          "role",
          "display_name",
          "phone_e164",
          "timezone",
          "companion_call_time",
          "companion_enabled",
          "created_at"
        ],
        "title": "MemberOut"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "OkOut"
      },
...
      "SimSayIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SimSayIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}